requirements.txt
django-autocomplete-light==3.12.1
django-bootstrap5==24.3
django-filter==24.3
Django==5.1.1
PyYAML==6.0.2
psycopg[binary]==3.2.9
